pyproject.toml
[project]
name = "cspyro-ae-2-5"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=0.6.0",
    "langchain-chroma>=0.2.0",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.10",
    "openai>=2.21.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.54.0",
    "tiktoken>=0.12.0",
    "youtube-transcript-api>=0.6.3",
    "yt-dlp>=2024.1.1",
]
